[Feat] Normalizing speed of monster
#14

~ 몬스터 이동 속도 정규화
~ 차이 식별용 이동속도 다른 몬스터 정보 추가 구성

+++ SlimeGG/Assets/Scripts/Monsters/Objects/MonsterAbilityStat.cs
@@ -3,5 +3,5 @@ public class MonsterAbilityStat : ElementStat
 {
     public int hp { get; set; }
     public int def { get; set; }
-    public int spd { get; set; }
+    public float spd { get; set; }
 }
